Bug 19016: Update other values
Here we have a problem!
If we assign a new id to the different rows that cannot be moved
safetly, we need to modify the other tables that do not have a foreign
key (for historical or laziness reasons).

For instance:
John is borrowernumber=42 and create a suggestion
(suggestion.suggestedby=42)
Jane is borrowernumber=42 in the deletedborrowers table. She created a
suggestion (same suggestedby value).

John will get a new id and the suggestion.suggestedby will not be
replaced without this patch. But with this fix, the 2 suggestions will
be marked as suggested by him.