WR237697: add support for a trusted HTTP header
This adds support for a 'trusted_header' option in koha-conf.xml that
specified an HTTP header that you trust that contains the userid. This
is to allow Koha to be behind a reverse proxy (for example, running
plack fronted by apache) that does user authentication/authorisation.

Note: for reasons I can't really tell, this doesn't work when apache is
running Koha as CGI, but does work under plack.

Test plan:
* have a koha-plack setup.
* configure apache to send it a header:
  RequestHeader set X_REMOTE_USER "testuserid"
* add <trusted_header>X_REMOTE_USER</trusted_header> to koha-conf.xml
* verify that the user behaves as though they're logged in.

Conflicts:
	C4/Auth.pm